+++ pkg/generic/ensure_release_label.go
@@ -25,19 +25,11 @@ func EnsureReleaseVersionLabel(ctx context.Context, ctrlClient client.Client, me
 			return nil, microerror.Maskf(errors.InvalidOperationError, "Object has no %s label, can't detect release version.", label.Cluster)
 		}
 
-		// Retrieve the `Cluster` CR related to this object.
-		cluster := &capiv1alpha3.Cluster{}
-		{
-			err := ctrlClient.Get(ctx, client.ObjectKey{Name: clusterID, Namespace: meta.GetNamespace()}, cluster)
-			if apierrors.IsNotFound(err) {
-				return nil, microerror.Maskf(errors.InvalidOperationError, "Looking for Cluster named %s but it was not found. Can't continue.", clusterID)
-			} else if err != nil {
-				return nil, microerror.Mask(err)
-			}
+		// Get release from Cluster.
+		release, err := getLabelValueFromCluster(ctx, ctrlClient, meta, label.ReleaseVersion)
+		if err != nil {
+			return nil, microerror.Mask(err)
 		}
-
-		// Extract release from Cluster.
-		release := cluster.GetLabels()[label.ReleaseVersion]
 		if release == "" {
 			return nil, microerror.Maskf(errors.InvalidOperationError, "Cluster %s did not have a release label set. Can't continue.", clusterID)
 		}
@@ -46,6 +38,29 @@ func EnsureReleaseVersionLabel(ctx context.Context, ctrlClient client.Client, me
 	}
 
 	return nil, nil
+}
+
+func getLabelValueFromCluster(ctx context.Context, ctrlClient client.Client, meta metav1.Object, labelName string) (string, error) {
+	clusterID := meta.GetLabels()[label.Cluster]
+	if clusterID == "" {
+		return "", microerror.Maskf(errors.InvalidOperationError, "Object has no %s label, can't detect cluster ID.", label.Cluster)
+	}
+
+	// Retrieve the `Cluster` CR related to this object.
+	cluster := &capiv1alpha3.Cluster{}
+	{
+		err := ctrlClient.Get(ctx, client.ObjectKey{Name: clusterID, Namespace: meta.GetNamespace()}, cluster)
+		if apierrors.IsNotFound(err) {
+			return "", microerror.Maskf(errors.InvalidOperationError, "Looking for Cluster named %s but it was not found. Can't continue.", clusterID)
+		} else if err != nil {
+			return "", microerror.Mask(err)
+		}
+	}
+
+	// Extract desired label from Cluster.
+	release := cluster.GetLabels()[labelName]
+
+	return release, nil
 }
 
 // Ensure the needed escapes are in place. See https://tools.ietf.org/html/rfc6901#section-3 .